Move Printer functionality into trait
